Fixed dashboard issue

+++ lib/screens/tabs/home/model/covid_bd_model.dart
@@ -14,7 +14,7 @@ class CovidBdData {
   int recovered;
   int active;
   int critical;
-  int casesPerOneMillion;
+  double casesPerOneMillion;
 
   CovidBdData({
     this.country,
